+++ src/main.rs
@@ -19,10 +19,8 @@ use std::error::Error;
 use std::path::PathBuf;
 use tokio::spawn;
 use tracing::{debug, info, error};
-use opentelemetry::global::{self};
-use telemetry::metrics_registry::{init_metrics_registry, get_metrics};
-use prism_telemetry::telemetry::{init_telemetry, build_resource, set_global_attributes};
-use prism_telemetry::logs::setup_log_subscriber;
+use crate::telemetry::metrics_registry::get_metrics;
+use crate::telemetry::init::init;
 
 pub static PRISM_MESSENGER_SERVICE_ID: &str = "prism_messenger";
 
@@ -53,35 +51,19 @@ async fn main() -> Result<(), Box<dyn Error>> {
         }
     };
 
+    // Initialize telemetry
     let mut attributes: Vec<(String, String)> = Vec::new();
-    attributes.extend(telemetry_config.global_labels.labels.clone().into_iter().map(|(k, v)| (k, v)));
-    attributes.push(("prism_host".to_string(), settings.prism.host.to_string() + ":" + &settings.prism.port.to_string()));
-
-    set_global_attributes(attributes.clone());
-
-    let resource = build_resource("prism-messenger-server".to_string(), attributes);
-
-    let (meter_provider, log_provider) = init_telemetry(&telemetry_config, resource).map_err(|e| anyhow::anyhow!(e.to_string()))?;
-
-    if let Some(ref provider) = meter_provider {
-        global::set_meter_provider(provider.clone());
-
-        // Initialize the metrics registry after setting the global meter provider
-        init_metrics_registry();
-    }
-
-    if let Some(ref provider) = log_provider {
-        // Initialize tracing subscriber
-        setup_log_subscriber(
-            telemetry_config.logs.enabled,
-            Some(provider)
-        );
-    }
+    attributes.push(("label".to_string(), "value".to_string()));
+    init(
+        &telemetry_config,
+        attributes,
+    )?;
 
     if let Some(metrics) = get_metrics() {
         metrics.record_node_info(
             vec![
                 ("version".to_string(), env!("CARGO_PKG_VERSION").to_string()),
+                ("prism_host".to_string(), settings.prism.host.to_string() + ":" + &settings.prism.port.to_string()),
             ]
         );
     }
@@ -96,7 +78,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     let webserver_task_handle = spawn(async move {
         debug!("starting webserver");
         if let Err(e) = webserver::start(&settings.webserver, context).await {
-            error!("Error occurred while running prover: {:?}", e);
+            error!("Error occurred while running webserver: {:?}", e);
         }
     });
