+++ src/telemetry/init.rs
@@ -1,0 +1,35 @@
+use crate::telemetry::metrics_registry::{init_metrics_registry, get_metrics};
+
+use opentelemetry::global::{self};
+use prism_telemetry::telemetry::{init_telemetry, build_resource, set_global_attributes};
+use prism_telemetry::logs::setup_log_subscriber;
+use prism_telemetry::config::TelemetryConfig;
+
+pub fn init(telemetry_config: &TelemetryConfig, attributes: Vec<(String, String)>) -> Result<(), anyhow::Error> {
+    // Initialize the telemetry system
+
+    let mut attributes = attributes.clone();
+    attributes.extend(telemetry_config.global_labels.labels.clone().into_iter().map(|(k, v)| (k, v)));
+
+    set_global_attributes(attributes.clone());
+
+    let resource = build_resource("prism-messenger-server".to_string(), attributes);
+
+    let (meter_provider, log_provider) = init_telemetry(&telemetry_config, resource).map_err(|e| anyhow::anyhow!(e.to_string()))?;
+
+    if let Some(ref provider) = meter_provider {
+        global::set_meter_provider(provider.clone());
+
+        // Initialize the metrics registry after setting the global meter provider
+        init_metrics_registry();
+    }
+
+    if let Some(ref provider) = log_provider {
+        // Initialize tracing subscriber
+        setup_log_subscriber(
+            telemetry_config.logs.enabled,
+            Some(provider)
+        );
+    }
+    Ok(())
+}
