+++ src/telemetry/metrics_registry.rs
@@ -8,32 +8,32 @@ use prism_telemetry::telemetry::build_attributes;
 
 // Struct to hold all metrics
 #[derive(Clone)]
-pub struct PrismMetrics {
+pub struct PrismMessengerMetrics {
     #[allow(dead_code)]
     meter: Meter,
     // Node info metric
     pub node_info: Gauge<u64>,
 }
 
-impl Default for PrismMetrics {
+impl Default for PrismMessengerMetrics {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl PrismMetrics {
+impl PrismMessengerMetrics {
     pub fn new() -> Self {
         info!("Initializing Prism metrics registry");
         let meter = global::meter("prism");
 
         let prefix = "prism_messenger_server_";
-
+        
         let node_info = meter
             .u64_gauge(format!("{}node_info", prefix))
             .with_description("Prism node info")
             .build();
 
-        PrismMetrics {
+        PrismMessengerMetrics {
             meter,
             node_info,
         }
@@ -50,25 +50,20 @@ impl PrismMetrics {
 
 // Global instance of PrismMetrics
 lazy_static! {
-    static ref METRICS: Mutex<Option<Arc<PrismMetrics>>> = Mutex::new(None);
+    static ref METRICS: Mutex<Option<Arc<PrismMessengerMetrics>>> = Mutex::new(None);
 }
 
 // Initialize the global metrics instance
 pub fn init_metrics_registry() {
     let mut metrics = METRICS.lock();
     if metrics.is_none() {
-        *metrics = Some(Arc::new(PrismMetrics::new()));
+        *metrics = Some(Arc::new(PrismMessengerMetrics::new()));
         info!("Prism metrics registry initialized");
     }
 }
 
 // Get a reference to the metrics registry
-pub fn get_metrics() -> Option<Arc<PrismMetrics>> {
-    match METRICS.try_lock() {
-        Some(guard) => guard.clone(),
-        None => {
-            tracing::warn!("Failed to acquire lock for metrics registry");
-            None
-        }
-    }
+pub fn get_metrics() -> Option<Arc<PrismMessengerMetrics>> {
+    let guard = METRICS.lock();
+    guard.clone()
 }
