+++ src/telemetry/mod.rs
@@ -1,1 +1,2 @@
+pub mod init;
 pub mod metrics_registry;
